refactor(agents): build agentbox claude from the command factory
The last and largest clone: `commands/claude.ts` (2,062 lines) becomes
`agents/claude/runtime.ts` (docker bindings, the credential-health sign-in
offers), `agents/claude/login-command.ts` (the headless print-URL / `--code`
protocol) and a descriptor whose five hooks are the reason the hook set exists
— `--plan` beside session teleport, the setup wizard, the plugin-native-deps
rebuild on create and on start, and the clipboard paste handlers.

Two orderings that looked like accidents are now declared, because both have a
reason: claude asks to sign in BEFORE its wizard can spend minutes re-baking a
stale base (`signInOfferTiming: 'before-gates'`), while codex and opencode ask
after the hub-routing decision so a box the control box will build never prompts
for a local login it won't use. Likewise `ensureInstalledOnCreate: false` keeps
claude from gaining a binary probe it has never run.

Golden CLI-surface fixture unchanged for all three agents. Whole-monorepo suite
green (4,224 tests).
