+++ apps/cli/src/agents/commands.ts
@@ -19,7 +19,7 @@
 import type { Command } from 'commander';
 import type { AttachOpenIn } from '@agentbox/config';
 import type { BoxRecord } from '@agentbox/sandbox-docker';
-import { claudeCommand, attachClaudeWrapped } from '../commands/claude.js';
+import { claudeCommand, attachClaudeWrapped } from './claude/command.js';
 import { codexCommand, attachCodexWrapped } from './codex/command.js';
 import { opencodeCommand, attachOpencodeWrapped } from './opencode/command.js';
